Created get_pruned_dict for streaming.
